Refactor punctuations into its own module
